Centralizing and DRY'ing up the codebase
So many parameters are centered around bill_date. Early on it made
testing easier just to specify it as an argument.

Upon review, bill_date is an omnipresent value. It exists everwhere
because it's so foundational to the context the code runs in.

So it bubbles up to the top along with it's cousin method days_in_month
where it can be easily accessed by any method.

A helper method in the specs make it just as easy to specify.